restructure and organize code better
