+++ src/components/Validator/LogsSection/index.tsx
@@ -1,0 +1,1 @@
+export * from './LogsSection'
