+++ src/components/Validator/LogsSection/types.ts
@@ -1,0 +1,13 @@
+import { LogEntry } from '@/schemas/k6'
+
+export type LogSource = 'browser' | 'runtime' | 'script'
+
+export interface ConsoleFilter {
+  levels: Array<LogEntry['level']>
+  sources: Array<LogSource>
+}
+
+export interface LogEntryWithSource {
+  source: LogSource
+  entry: LogEntry
+}
